requirements.txt
annotated-doc==0.0.4
asgiref==3.11.1
click==8.3.1
Django==5.2.1
django-htmx==1.27.0
django-tailwind-cli==4.5.1
django-typer==3.6.2
django-unfold==0.80.2
markdown-it-py==4.0.0
mdurl==0.1.2
Pygments==2.19.2
rich==14.3.3
semver==3.0.4
shellingham==1.5.4
sqlparse==0.5.5
typer==0.24.0
whitenoise==6.11.0
